+++ src/main/java/net/kyrptonaught/customportalapi/PerWorldPortals.java
@@ -1,0 +1,24 @@
+package net.kyrptonaught.customportalapi;
+
+import net.kyrptonaught.customportalapi.util.PortalLink;
+import net.minecraft.block.Block;
+import net.minecraft.util.registry.Registry;
+
+import java.util.HashSet;
+
+public class PerWorldPortals {
+    private static final HashSet<Block> worldPortals = new HashSet<>();
+
+    public static void removeOldPortalsFromRegistry() {
+        for (Block block : worldPortals) {
+            CustomPortalApiRegistry.portals.remove(block);
+        }
+        worldPortals.clear();
+    }
+
+    public static void registerWorldPortal(PortalLink portalLink) {
+        Block blockId = Registry.BLOCK.get(portalLink.block);
+        worldPortals.add(blockId);
+        CustomPortalApiRegistry.addPortal(blockId, portalLink);
+    }
+}