+++ src/main/java/net/kyrptonaught/customportalapi/mixin/InGameHudMixin.java
@@ -5,7 +5,7 @@ import net.kyrptonaught.customportalapi.CustomPortalApiRegistry;
 import net.kyrptonaught.customportalapi.CustomPortalBlock;
 import net.kyrptonaught.customportalapi.CustomPortalsMod;
 import net.kyrptonaught.customportalapi.util.ColorUtil;
-import net.kyrptonaught.customportalapi.util.PlayerInCustomPortal;
+import net.kyrptonaught.customportalapi.util.EntityInCustomPortal;
 import net.kyrptonaught.customportalapi.util.PortalLink;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.Blocks;
@@ -29,7 +29,7 @@ public class InGameHudMixin {
 
     @Redirect(method = "renderPortalOverlay", at = @At(value = "INVOKE", target = "Lcom/mojang/blaze3d/systems/RenderSystem;color4f(FFFF)V"))
     public void changeColor(float red, float green, float blue, float alpha) {
-        if (((PlayerInCustomPortal) client.player).getTimeInPortal() > 0) {
+        if (((EntityInCustomPortal) client.player).getTimeInPortal() > 0) {
             PortalLink link = CustomPortalApiRegistry.portals.get(CustomPortalBlock.getPortalBase(client.world, client.player.getBlockPos()));
             float[] colors = link != null ? ColorUtil.getColorForBlock(link.colorID) : DyeColor.WHITE.getColorComponents();
             RenderSystem.color4f(colors[0], colors[1], colors[2], alpha);
@@ -39,7 +39,7 @@ public class InGameHudMixin {
 
     @Redirect(method = "renderPortalOverlay", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/render/block/BlockModels;getSprite(Lnet/minecraft/block/BlockState;)Lnet/minecraft/client/texture/Sprite;"))
     public Sprite renderCustomPortalOverlay(BlockModels blockModels, BlockState blockState) {
-        if (((PlayerInCustomPortal) client.player).getTimeInPortal() > 0) {
+        if (((EntityInCustomPortal) client.player).getTimeInPortal() > 0) {
             return blockModels.getSprite(CustomPortalsMod.portalBlock.getDefaultState());
         }
         return blockModels.getSprite(Blocks.NETHER_PORTAL.getDefaultState());